+++ crates/bevy_ecs/src/world/identifier.rs
@@ -32,7 +32,7 @@ mod tests {
 
     #[test]
     fn world_ids_unique() {
-        let ids = std::iter::repeat_with(WorldId::new)
+        let ids = core::iter::repeat_with(WorldId::new)
             .take(50)
             .map(Option::unwrap)
             .collect::<Vec<_>>();
@@ -49,7 +49,7 @@ mod tests {
     // #[should_panic]
     // fn panic_on_overflow() {
     //     MAX_WORLD_ID.store(usize::MAX - 50, Ordering::Relaxed);
-    //     std::iter::repeat_with(WorldId::new)
+    //     core::iter::repeat_with(WorldId::new)
     //         .take(500)
     //         .for_each(|_| ());
     // }