+++ crates/bevy_ecs/src/world/mod.rs
@@ -692,7 +692,7 @@ impl World {
 
     /// Returns an iterator of entities that had components of type `T` removed
     /// since the last call to [`World::clear_trackers`].
-    pub fn removed<T: Component>(&self) -> std::iter::Cloned<core::slice::Iter<'_, Entity>> {
+    pub fn removed<T: Component>(&self) -> core::iter::Cloned<core::slice::Iter<'_, Entity>> {
         if let Some(component_id) = self.components.get_id(TypeId::of::<T>()) {
             self.removed_with_id(component_id)
         } else {
@@ -705,7 +705,7 @@ impl World {
     pub fn removed_with_id(
         &self,
         component_id: ComponentId,
-    ) -> std::iter::Cloned<core::slice::Iter<'_, Entity>> {
+    ) -> core::iter::Cloned<core::slice::Iter<'_, Entity>> {
         if let Some(removed) = self.removed_components.get(component_id) {
             removed.iter().cloned()
         } else {
@@ -1227,13 +1227,13 @@ impl World {
     /// Sends an [`Event`](crate::event::Event).
     #[inline]
     pub fn send_event<E: crate::event::Event>(&mut self, event: E) {
-        self.send_event_batch(std::iter::once(event));
+        self.send_event_batch(core::iter::once(event));
     }
 
     /// Sends the default value of the [`Event`](crate::event::Event) of type `E`.
     #[inline]
     pub fn send_event_default<E: crate::event::Event + Default>(&mut self) {
-        self.send_event_batch(std::iter::once(E::default()));
+        self.send_event_batch(core::iter::once(E::default()));
     }
 
     /// Sends a batch of [`Event`](crate::event::Event)s from an iterator.