+++ crates/bevy_ecs/src/world/world_cell.rs
@@ -324,13 +324,13 @@ impl<'w> WorldCell<'w> {
     /// Sends an [`Event`](crate::event::Event).
     #[inline]
     pub fn send_event<E: Event>(&self, event: E) {
-        self.send_event_batch(std::iter::once(event));
+        self.send_event_batch(core::iter::once(event));
     }
 
     /// Sends the default value of the [`Event`](crate::event::Event) of type `E`.
     #[inline]
     pub fn send_event_default<E: Event + Default>(&self) {
-        self.send_event_batch(std::iter::once(E::default()));
+        self.send_event_batch(core::iter::once(E::default()));
     }
 
     /// Sends a batch of [`Event`](crate::event::Event)s from an iterator.