+++ crates/bevy_scene/src/scene_loader.rs
@@ -1,3 +1,4 @@
+use alloc::boxed::Box;
 use crate::serde::SceneDeserializer;
 use anyhow::Result;
 use bevy_app::AppTypeRegistry;