+++ examples/stress_tests/transform_hierarchy.rs
@@ -440,7 +440,7 @@ fn gen_tree(depth: u32, branch_width: u32) -> Vec<usize> {
     // the tree is built using this pattern:
     // 0, 0, 0, ... 1, 1, 1, ... 2, 2, 2, ... (count - 1)
     (0..count)
-        .flat_map(|i| std::iter::repeat(i).take(branch_width.try_into().unwrap()))
+        .flat_map(|i| core::iter::repeat(i).take(branch_width.try_into().unwrap()))
         .collect()
 }